requirements.txt
attrs==22.2.0
aws-cdk.assets==1.198.1
aws-cdk.aws-acmpca==1.198.1
aws-cdk.aws-apigateway==1.198.1
aws-cdk.aws-applicationautoscaling==1.198.1
aws-cdk.aws-autoscaling-common==1.198.1
aws-cdk.aws-certificatemanager==1.198.1
aws-cdk.aws-cloudformation==1.198.1
aws-cdk.aws-cloudwatch==1.198.1
aws-cdk.aws-codeguruprofiler==1.198.1
aws-cdk.aws-codestarnotifications==1.198.1
aws-cdk.aws-cognito==1.198.1
aws-cdk.aws-ec2==1.198.1
aws-cdk.aws-ecr==1.198.1
aws-cdk.aws-ecr-assets==1.198.1
aws-cdk.aws-efs==1.198.1
aws-cdk.aws-elasticloadbalancingv2==1.198.1
aws-cdk.aws-events==1.198.1
aws-cdk.aws-iam==1.198.1
aws-cdk.aws-kms==1.198.1
aws-cdk.aws-lambda==1.198.1
aws-cdk.aws-logs==1.198.1
aws-cdk.aws-route53==1.198.1
aws-cdk.aws-s3==1.198.1
aws-cdk.aws-s3-assets==1.198.1
aws-cdk.aws-sagemaker==1.198.1
aws-cdk.aws-signer==1.198.1
aws-cdk.aws-sns==1.198.1
aws-cdk.aws-sqs==1.198.1
aws-cdk.aws-ssm==1.198.1
aws-cdk.aws-stepfunctions==1.198.1
aws-cdk.cloud-assembly-schema==1.198.1
aws-cdk.core==1.198.1
aws-cdk.custom-resources==1.198.1
aws-cdk.cx-api==1.198.1
aws-cdk.region-info==1.198.1
cattrs==22.2.0
constructs==3.4.293
exceptiongroup==1.1.1
importlib-resources==5.12.0
jsii==1.80.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.5.0
zipp==3.15.0
